Add a function to convert a shape to a spacing
The `shape_to_spacing` function was ported from Verde. Some warnings
were fixed and added input validation and tests for n-dimensional grids.
Created the public `check_shape` function and used it in
`grid_coordinates` as well. Moved the input validations functions to
their own module. Moved `spacing_to_size` to the new `_utils` module
that also has `shape_to_spacing`.
