+++ src/bordado/__init__.py
@@ -8,12 +8,13 @@
 These are the functions and classes that make up the Bordado API.
 """
 
-from ._coordinates import check_coordinates
 from ._grid import grid_coordinates
-from ._line import line_coordinates, spacing_to_size
+from ._line import line_coordinates
 from ._random import random_coordinates
-from ._region import check_region, get_region, inside, pad_region
+from ._region import get_region, inside, pad_region
 from ._split import block_split, expanding_window, rolling_window
+from ._utils import shape_to_spacing, spacing_to_size
+from ._validation import check_coordinates, check_region, check_shape
 from ._version import __version__
 
 # Append a leading "v" to the generated version by setuptools_scm
